Added exploded head
Head will explode into multiple pieces when hit like glass smashing. This is done with the explode head scrip that will spawn a prefab of a shattered version of the head over the normal head. Later there will be a round script that will show the player has won when the head has exploded