Added more Scenary to the game and added explosion of heads
Floating islands with treses on them have been added to the scene to add a bit more depth to the game, while towers have been reduced in size to make it easier for players to see other towers near them without having to walk too much. An explosion script was added to the players heads so when they are shot by and arrow the head of glass will shatter. Made sure to turn off the collider on the arrow until it is fired when it will be turned back on. This was because of a bug where heads could explode when players pulled a arrow out of there quiver. More work done on the round over script with some models being made for showing the players they have won or lost the match